[feat][client][ssamc] barcode2菜单搜索

+++ src/client/DevkitPrism/Modules/Devkit.Modules.Ssamc/DB/Entities/SYS_MENU.cs
@@ -144,7 +144,7 @@ public partial class SYS_MENU
     /// 删除标记
     /// </summary>
     [Column(TypeName = "NUMBER(1)")]
-    public int? IS_DELETE { get; set; }
+    public short? IS_DELETE { get; set; }
 
     /// <summary>
     /// 处理标志
